requirements.txt
Flask==2.3.3
spacy==3.6.1
pandas==2.0.3
scikit-learn==1.3.0
numpy==1.24.2
nltk==3.8.1
PyPDF2==3.0.1
python-docx==0.8.11
